…up by columns

<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
Problem Summary:
Right now for count-distinct agg we can only execute the dedup part on TiFlash and gather all the intermediate data to get final result on TiDB. However the execution in TiDB will be the performance bottleneck so we hope to push down this part in MPP mode.

### What is changed and how it works?
What's Changed:
We check the group by columns and distinct functions during `exhaustPhysicalPlans` and set the physical hint to `MppScalar` indicating this is a scalar function  then we will push down all the calculation. This tune only takes effect in the MPP mode. 

How it Works:
for a simple statement: `select count(distinct a) from t;`. The performace advance is great:
before tune:
```
mysql> explain analyze select count(distinct a) from t;
+------------------------------+-------------+----------+-------------------+---------------+-----------------------------------------------------------------------------------------------------------------+------------------------------------------+----------+------+
| id                           | estRows     | actRows  | task              | access object | execution info                                                                                                  | operator info                            | memory   | disk |
+------------------------------+-------------+----------+-------------------+---------------+-----------------------------------------------------------------------------------------------------------------+------------------------------------------+----------+------+
| HashAgg_13                   | 1.00        | 1        | root              |               | time:18.2s, loops:2                                                                                             | funcs:count(distinct test.t.a)->Column#4 | 932.1 MB | N/A  |
| └─TableReader_15             | 1.00        | 30950159 | root              |               | time:461.5ms, loops:30593, cop_task: {num: 640, max: 0s, min: 0s, avg: 0s, p95: 0s, copr_cache_hit_ratio: 0.00} | data:ExchangeSender_14                   | N/A      | N/A  |
|   └─ExchangeSender_14        | 1.00        | 30950159 | batchCop[tiflash] |               | tiflash_task:{proc max:1.11s, min:0s, p80:1ms, p95:8ms, iters:768, tasks:640, threads:48}                       | ExchangeType: PassThrough                | N/A      | N/A  |
|     └─HashAgg_6              | 1.00        | 30950159 | batchCop[tiflash] |               | tiflash_task:{proc max:1.11s, min:0s, p80:1ms, p95:8ms, iters:768, tasks:640, threads:3}                        | group by:test.t.a,                       | N/A      | N/A  |
|       └─TableFullScan_12     | 64349746.00 | 64238857 | batchCop[tiflash] | table:t       | tiflash_task:{proc max:274.6ms, min:0s, p80:0s, p95:0s, iters:1044, tasks:640, threads:48}                      | keep order:false                         | N/A      | N/A  |
+------------------------------+-------------+----------+-------------------+---------------+-----------------------------------------------------------------------------------------------------------------+------------------------------------------+----------+------+
5 rows in set (18.22 sec)
```

after tune
```
mysql> explain analyze select count(distinct a) from t;
+----------------------------------+----------+----------+-------------------+---------------+--------------------------------------------------------------------------------------------+------------------------------------------+--------+------+
| id                               | estRows  | actRows  | task              | access object | execution info                                                                             | operator info                            | memory | disk |
+----------------------------------+----------+----------+-------------------+---------------+--------------------------------------------------------------------------------------------+------------------------------------------+--------+------+
| TableReader_18                   | 1.00     | 1        | root              |               | time:3.29s, loops:2, cop_task: {num: 1, max: 0s, proc_keys: 0, copr_cache_hit_ratio: 0.00} | data:ExchangeSender_17                   | N/A    | N/A  |
| └─ExchangeSender_17              | 1.00     | 1        | batchCop[tiflash] |               | tiflash_task:{time:3.19s, loops:1, threads:20}                                             | ExchangeType: PassThrough                | N/A    | N/A  |
|   └─HashAgg_14                   | 1.00     | 1        | batchCop[tiflash] |               | tiflash_task:{time:3.19s, loops:1, threads:1}                                              | funcs:count(distinct test.t.a)->Column#4 | N/A    | N/A  |
|     └─ExchangeReceiver_16        | 1.00     | 30950159 | batchCop[tiflash] |               | tiflash_task:{time:1.09s, loops:640, threads:20}                                           |                                          | N/A    | N/A  |
|       └─ExchangeSender_15        | 1.00     | 30950159 | batchCop[tiflash] |               | tiflash_task:{time:1.23s, loops:768, threads:48}                                           | ExchangeType: PassThrough                | N/A    | N/A  |
|         └─HashAgg_6              | 1.00     | 30950159 | batchCop[tiflash] |               | tiflash_task:{time:1.23s, loops:768, threads:3}                                            | group by:test.t.a,                       | N/A    | N/A  |
|           └─TableFullScan_12     | 10000.00 | 64238857 | batchCop[tiflash] | table:t       | tiflash_task:{time:256.9ms, loops:1044, threads:48}                                        | keep order:false, stats:pseudo           | N/A    | N/A  |
+----------------------------------+----------+----------+-------------------+---------------+--------------------------------------------------------------------------------------------+------------------------------------------+--------+------+
7 rows in set (6.24 sec)
```


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Manual test (add detailed scripts or steps below)


### Release note <!-- bugfixes or new feature need a release note -->

- planner/core: thoroughly push down count-distinct agg in the MPP mode.